challenge1/requirements.txt
numpy
pandas
scikit-learn
Pillow
tqdm
torch
torchvision
efficientnet_pytorch
